0 error 0 warnings on Windows
Silence all first-party compiler warnings in Debug and Release
on MSVC (VS 2019 16.10) so the build matches the AppleClang
target verified in b8bf64d.

Build system (engine/CMakeLists.txt):
- Verify /external:W0 + /external:I silences vendored headers
  (the b8bf64d note said UNTESTED on Windows).
- Add /wd4127 /wd4245 globally. Both warnings escape ImGui +
  ImGuizmo TU files (not headers) that we compile directly:
  C4127 fires on imgui_impl_opengl3.cpp's IM_ASSERT(0 && ...);
  C4245 fires on ImGuizmo's int -> ImGuiID casts. Neither fires
  in engine code.

Engine code fixes (mirrors the b8bf64d style):
- C4100 unreferenced formal params: rename to /*name*/ for
  virtual-stub overrides that need the name visible (Component.h
  Load/Save); use (void)param; in plain definitions where the
  param is purely positional (Component.cpp OnDetaching/
  OnOwnerDestructing, AnimationPlayer.cpp OnOwnerDestructing,
  EditorNodeProperties.cpp's six Render*Body helpers,
  Editor3DPreview.cpp ComputeViewProj, EditorWindows.cpp
  RenderAssetTile).
- C4456 declaration hides previous local: rename key/btn in
  Engine.cpp's SFML event dispatch (each else-if branch in its
  own scope with keyPressed/keyReleased etc.) and p1..p4 in
  EditorWindows.cpp's Uniform1f..4f dispatch (preserves the
  if/else-if short-circuit; a sequence of independent ifs would
  have lost it).
- C4244/C4267 size_t->int / size_t->unsigned / size_t->rapidjson::SizeType
  truncation: widen loop counters to size_t where they index a
  container (AnimationClip, AnimationUtil, ArrayBuffers), and add
  static_cast<unsigned int> / static_cast<GLsizei> /
  static_cast<GLuint> / static_cast<GLint> / static_cast<int> at
  the specific call sites that pass a size_t to a fixed-width API
  (SaveArray, glDrawElements, glBindTexture, glUniform*, BindMatrices,
  rapidjson::Key/String). Component.cpp::Pass.cpp::RenderUtil.cpp
  etc.
- C4701 potentially uninitialized: init Texture.cpp width/height
  to 0.
- C4457 declaration hides function parameter: rename local
  color[] -> border[] in Texture::SetBorderColor.
- C4245 'return': -1 -> 0xFFFFFFFFu sentinel for Pass::GetTextureIndex
  with matching check in Pass::GetTexture.
- C4389 signed/unsigned mismatch in ArrayBuffers size_t sizeNew.
- C4005 macro redefinition: #ifndef guard around WIN32_LEAN_AND_MEAN
  in FbxConverter.cpp.

Encoding (Transform.cpp + 74 source files):
- C4828 fired on GBK-encoded em-dashes/dashes/etc once /utf-8
  was on. Sweep replaced common non-ASCII punctuation across
  the engine (--, *, deg, ->, ...), translated the only CJK
  comment (Transform.cpp TODO), and converted files from GBK to
  UTF-8 where needed. Light.h keeps a '45 deg' inline note,
  EditorAssetWindows.h has '128x128', etc.

Vendored-header workarounds (NOT pragma suppression):
- FileUtil.cpp + Texture.cpp wrap #include "stb_image.h" with
  #pragma warning(push/disable: 4100 4312 4456 4505/pop) inside
  #if PLATFORM_WINDOWS. stb_image.h has no internal pragma, so
  the outer pair survives.
- Gui.cpp / Editor.cpp / EditorWindows.cpp do NOT wrap ImGui +
  ImGuizmo includes. imgui.h has its own
  #pragma warning(push) at line 135 / #pragma warning(pop) at
  line 4506; an outer pair would be consumed by imgui's inner
  pop and trigger C4193. The escaping C4127 / C4245 / C4267 from
  ImGui/ImGuizmo TU files are silenced globally via /wd4127
  /wd4245 (in CMakeLists.txt) rather than at include sites.

ThreadUtil / Log.h:
- C4251 'STL in exported class' on ThreadUtil (std::thread::id,
  std::unordered_map, std::thread, std::mutex, std::condition_variable)
  and Log::Record (std::string, std::stringstream). Wrap each
  FURY_API class body in
  #pragma warning(push/disable: 4251/pop). Informational only --
  all consumers build with the same MSVC STL.

Casts for conversions the AppleClang build did not need:
- Pass.cpp::GetTextureIndex returns 0xFFFFFFFFu (not -1 cast) and
  the caller compares against 0xFFFFFFFFu.
- Shader.cpp::BindTexture casts size_t textureId -> GLuint at
  the glBindTexture call and the m_TextureID - GL_TEXTURE0
  subtraction to GLint at the glUniform1i call.
- ThreadUtil::ThreadUtil takes size_t (was unsigned int) so
  Singleton<ThreadUtil, size_t>::Initialize / make_shared no
  longer trips xutility's size_t->unsigned narrowing.

Verified: Debug + Release builds produce fury.exe + furye.exe
under examples/{Debug,Release}/. Only warnings remaining are
three D9025s from SFML's vendored third-party deps (SheenBidi,
FreeType, HarfBuzz) -- those are SFML's CMakeLists, not ours.

+++ engine/Fury/AnimationClip.cpp
@@ -137,15 +137,15 @@ namespace fury
 
 		for (auto channel : m_Channels)
 		{
-			int posCount = channel->positions.size();
+			size_t posCount = channel->positions.size();
 			if (posCount > 0)
 				Try(channel->positions[posCount - 1].tick);
 
-			int rotCount = channel->rotations.size();
+			size_t rotCount = channel->rotations.size();
 			if (rotCount > 0)
 				Try(channel->rotations[rotCount - 1].tick);
 
-			int sclCount = channel->scalings.size();
+			size_t sclCount = channel->scalings.size();
 			if (sclCount > 0)
 				Try(channel->scalings[sclCount - 1].tick);
 		}
@@ -197,7 +197,7 @@ namespace fury
 
 	int AnimationClip::GetChannelCount() const
 	{
-		return m_Channels.size();
+		return static_cast<int>(m_Channels.size());
 	}
 
 	AnimationClip::ChannelPtr AnimationClip::AddChannel(const std::string &name)
